Improve checking for modified files
https://review.openstack.org/#/c/48726/ introduced a typo in the
filename of an include directive that was not detected by our
validation scripts since we did only check whether a file
included a modified file but omit checking pom.xml for
modification of included files. So, if bk-* gets modified,
we do not notice it and therefore won't build the book.

With this change, the books will be build. I also added
a couple of comments.

Change-Id: Idba71ac4f9f1e4de2c414bcc0110351966df3248