Rework repository to make it an installable python package
This is an overhaul of the complete content to make it
a separate python package that can be installed on the jenkins
machines as well as on editor's machines.

The goal of this patchset is to package everything
and get the command "openstack-doc-test" running so that
it can be used for gating.

This will need further refinement for the other commands.

Change-Id: Icc2f1807dd5ab5fb2f83c05d1b3895b3a9a0dbaf