Reworked autohelp to support dup option names in different groups
autohelp.py can't currently distinguish between options with the same
name in different groups, leading to rather poor DocBook output. This
commit changes the flagmappings files to have the group name prepended
with a dot if the group name isn't DEFAULT.

This commit rewrites much of autohelp.py, simplifies the command line,
deletes a ton of code, and reuses small portions of the sample config
file generator for oslo-incubator. In a happier world, this tool and
the config file generator could share more of the module importing
and option munging code.

This commit does not add group information to the DocBook, even though
it's now available in the code, along with the importing module name.
I want to minimize differences in DocBook output for testing, then add
more info to the DocBook.

Change-Id: I1b12659a80fa7989890c8b9b855c9130b85b7f3b