Allow setting of publish directory
You can now set the publication directory for a book. This will
allow to make more generic Jenkins jobs since we can now say for example
that api-quick-start-onepager-external gets published as
api-quick-start.

Fix publishing so that books published to trunk or havana etc are
copied to the proper subdirectory of publish-docs.

Change-Id: Ifb45a352e6f803321dcfcfcc5f8f17b7876f209b