Fix error reporting of doctest and parallel builds
Fix error handling in openstack-doc-test, now really abort if an
error occurs. The call to write in the exception was wrong and masked
the exception.

OpenStack infra changed the setup so that we now always destroy a node
after building - and thus have a clean node each time.
Now, when mvn is called, it needs to download lots of files into the
.ml2 repository (previously this was done already).
Now, the cripts start mvn jobs in parallel - and then we get a race
between these downloads and corruption of the .m2 directory.
Avoid this race with waiting for the first build job to finish before
starting further builds.

Add options --parallel/--noparallel to disable parallel building
completely.

Improve debug output.

Related-Bug: #1277082

Change-Id: Ifa62e0bb5b4f655b0d36b9e16bb0bb8adcd4cbd5