Place </screen> on proper line, fix swift usage
Change way on how the screens are generated so that we really
have </screen> on the closing line.

Also, update swift usage line so that it is not inside a screen but in a
normal para.

Add new option --output-directory for defining the directory where the
files will be written to.

Change-Id: Ie5d039ad673290720993586bd137c51183afa822