extract_swift_flags.py: implement a 'dump' argument
This argument dumps the list of swift config options discovered by the
script. The options are stored as oslo.config options and the dump
format is the same as the autohelp dump format. The goal is to support
swift in the diff_branches.py script.

Part of the script as been refactored to first store the options, then
write/dump them, instead of doing both at the same time.

Change-Id: Ie11f0eb352c79b7a5126599adb96ac1307f2956f