Improve the output format for option names
The diff_branches script stores the location of options as a simple
string, which differs from what the user actually see in the config
file. This patch improves the output of the option name depending on its
format in the 'dump' output.

We add brackets around the [section_name], and handle the file names for
swift options.

Change-Id: I03e12fc04f1e6f2f0831534cdff6dac048d76d92