Added support for *-manage CLI doc generation
Support for the *-manage commands was added to commands.py. Two new
options were added as well, --all-api and --all-manage.

--all-api generates files for all API commands
--all-manage generates files for all manage commands

--all generates files for all commands, API and manage

At present, only trove-manage is supported for *-manage commands.

As an aside, the Swift CLI help command is different from most other
CLI help, and as it is the same as the *-manage help it was changed to
use the same help syntax.

Change-Id: I851f39a405f91216257a5965a6ce5c01d1889851
Closes-Bug: #1338806