Autohelp: sanitize the value during discovery
Most of the default value sanitization happened during the docbook
writting, and not during the 'dump' of the options. This resulted in
invalid values in the xml files produced by diff_branches.py.

This patch moves all the sanitization code to the same function,
ensuring that the sanitization happens during options discovery.

Change-Id: I4d7e5d601baac5294937525ede8655da6b25e6cf