Change config changes generator script
Change diff_branches.py to generate the config-changes file with an
xml:id that always contains the release name instead of the name of
the new branch. This way it can be avoided that the id changes, when
stable branch is cut close to the end of a release cycle.

Change-Id: I96ef56fef7dac98f344296d1217586326740b46e