Add final newline to file formatting
The JSON formatting created by jsoncheck.py omits the file's final
newline. It claims that files with such a newline need reformatting
and removes the final newline in a file when reformatting.

This patch brings jsoncheck.py in line with the rest of the tools
which prefer text files to end with a newline.

Update release notes for pushing out the bug fix.

Closes-Bug: #1403159
Change-Id: I35eab544b183c05b167ce9917d92f0f1f919b745