Fix xml tagging for non-swift config tables
A previous patch attempted to introduce semantic tagging for
options and their default values in the automatically
generated configuration reference tables.

While this worked well for swift, incorrect code was used
for the XML generation of the main autohelp.py method.

This patch fixes the string based code to use XML subElements
as was already done for extract_swift_flags.

Change-Id: I93a89b726d66acd0778e8365f47d5ee2f2e0b3c5