Increase visibility of the warning in autodoc files
Autogenerate-config-docs creates files that live in the
openstack-manuals repository, but are automatically generated
and should not be edited.

On occasion, contributors miss the warning not to edit the files.

This patch increases the previous warning so that it should take up
enough screen real-estate in a text editor that it would be less
easily skipped, and adds some additional information on what to do
instead.

Change-Id: I800e1ed19427e00f903046731fe8f22f22e4c6db