diff_branches: use jinja templates to create the files
Generating docbook using lxml works fine, but the switch to RST tables
will make the code more diffiicult to read and maintain. Using jinja
templates will shorten the code and avoid dealing with complicated code.

This patch implements the use of jinja for docbook tables, and will ease
the implementation of RST output.

Change-Id: Ia5a75ced773b7e6e8faa7f89cb7620d986c47718