Move the config items in the manuals repository
Modifying configuration options (ignored modules, extra repositories,
hooks) for autohelp requires an update of the doc-tools repository,
which doesn't make much sense since the goal is to update the manuals
repository.

This patch removes the configuration options from the repository and
updates the scripts to use the configuration items in openstack-manuals.

Change-Id: I630d2a2fea690f6360039aa100331678468d18dc