Better identify deprecated configuration options
If a configuration option is tagged as deprecated in the code and the
help text does not start with DEPRECATED: add this to better identify
options that are no longer used.

If the code provides an additional deprecated reason, append this to
the provided help text of the configuration option.

Change-Id: I1a7c76c0109a6e59cc87d0ddc942c6d1cfb9dc91