Remove parallel building
Parallel building is not working reliable with errors, I just had a
problem were the script run successfully but should have failed.
Let's be on the safe side...

Also fix  "WARNING: term not in glossary: fqdn" and
"WARNING: unknown document: example-architectures".

This shows that the parallel building hit problems, otherwise this
would not show up now.

Change-Id: I731de4def0a00e8f94c0d997d5bd66f5ec438e87
Co-Authored-By: Andreas Jaeger <aj@suse.com>