Include 'deprecation' as a flag
There is support for displaying deprecation information in a somewhat
structured manner (proper lists, emphasis on titles, etc.). However, the
deprecation information is still included in the help text rather than
in this flags section. This makes things more difficult than they need
to be.

Resolve this by instead including any deprecation information in the
flags section. This will transform output from something like this:

  (String) DEPRECATED: Host to locate redis. Replaced by
  [DEFAULT]/transport_url

to this:

  (String) Host to locate redis.

  Deprecated
    Replaced by [DEFAULT]/transport_url

Which should be a little more obvious.

It's necessary to fix the template, which added the flags at the same
level as the title and description, resulting in varying numbers of
bullet points for the list.

Change-Id: Ifaee90a814fa3efee8b41eb0411535cff70e7309