Make flagmapping optional
Support for flagmappings was useful in the days before OptGroup was
available. Today, it's mostly a burden and not necessary for projects
such as nova. Make this feature optional, allowing us to make use of
built-in oslo.config where available.

The actual flagmapping files themselves are found in another project and
cannot be removed here. This will be done in a separate change.

Change-Id: I27296954ad8ce2010fc464ad81c8da4859ba2531