Remove "cleanup" tools
There were a couple of scripts here:

- prettify.py
- remove_trailing_whitespaces.sh
- remove_unnecessary_spaces.py
- retf

Most of them are self-explanatory with the exception of 'retf', which I
don't understand but which looks untouched in years and therefore
useless. For the other three, we have tooling in place to ensure these
"issues" don't get into source control and, even if they did, there's
questionable value in rewriting stuff just to remove excess whitespace.
Kill them all.

Change-Id: Ifbd2d50a669251484056552da21d23906f8d6178